requirements.txt
django
django-paypal
WeasyPrint